Move `fj::Sweep` to dedicated module
