+++ crates/fj/src/lib.rs
@@ -24,10 +24,12 @@ mod angle;
 mod group;
 mod shape_2d;
 mod shape_3d;
+mod sweep;
 mod transform;
 
 pub use self::{
-    angle::*, group::Group, shape_2d::*, shape_3d::*, transform::Transform,
+    angle::*, group::Group, shape_2d::*, shape_3d::*, sweep::Sweep,
+    transform::Transform,
 };
 pub use fj_proc::*;
 #[cfg(feature = "serde")]
